fix: tx direction
